+++ src/tests/RuleSets/Rules/TypeScript/index.ts
@@ -14,6 +14,7 @@ import { FuncCallSpacing } from "./FuncCallSpacing";
 import { MemberDelimiterStyle } from "./MemberDelimiterStyle";
 import { MemberOrdering } from "./MemberOrdering";
 import { NamingConvention } from "./NamingConvention";
+import { NoConfusingVoidExpression } from "./NoConfusingVoidExpression";
 import { NoDynamicDelete } from "./NoDynamicDelete";
 import { NoEmptyFunction } from "./NoEmptyFunction";
 import { NoEmptyInterface } from "./NoEmptyInterface";
@@ -61,6 +62,7 @@ export let TypeScriptTests = new ContainerSuite(
         MemberDelimiterStyle,
         MemberOrdering,
         NamingConvention,
+        NoConfusingVoidExpression,
         NoDynamicDelete,
         NoEmptyFunction,
         NoEmptyInterface,